Highlight empty widget areas when dragging within
