+++ packages/edit-widgets/src/blocks/widget-area/editor.scss
@@ -27,15 +27,16 @@
 	}
 }
 
-// Add some spacing above the inner blocks so that the block toolbar doesn't
-// overlap the panel header.
-.wp-block-widget-area > .components-panel__body > div > .editor-styles-wrapper > .block-editor-inner-blocks {
-	padding-top: $grid-unit-30;
+.edit-widgets-highlight-drop-zone {
+	outline: 2px dashed var(--wp-admin-theme-color);
+}
 
+.wp-block-widget-area > .components-panel__body > div > .editor-styles-wrapper > .block-editor-inner-blocks > .block-editor-block-list__layout {
 	// Ensure the widget area block lists have a minimum height so that it doesn't
 	// collapse to zero height when it has no blocks. When that happens the block
 	// can't be used as a drop target.
-	> .block-editor-block-list__layout {
-		min-height: $grid-unit-40;
-	}
+	min-height: $grid-unit-40;
+	// Add some spacing above the inner blocks so that the block toolbar doesn't
+	// overlap the panel header.
+	margin-top: $grid-unit-30;
 }
